Set the expiration start timestamp if required
Currently when a message has sent an expire timer, it will be instantly
deleted upon receipt on Signal Android as the expiration start timestamp
has no value. This PR sets that value if required to the timestamp the
message was sent at.

I don't think something similar is required for non-sync messages, but I
have not yet tested this.

Somewhat related to https://github.com/whisperfish/presage/issues/93.